Move cursor `Transformation` to `mouse::cursor` module
